feat: Enhance security configuration and update mapper methods
(cherry picked from commit b0397a542a8969ba059a3e222353c092a2b76522)

+++ src/main/java/com/mdv/appstore/mapper/ProductMapper.java
@@ -21,9 +21,9 @@ public interface ProductMapper {
 
     void updateTotalQuantity(@Param("id") Long id, @Param("totalQuantity") Long totalQuantity);
 
-    void increaseSoldQuantity(@Param("id") Long id, @Param("quantity") Long quantity);
+    void increaseSoldQuantity(@Param("id") Long id, @Param("quantity") Integer quantity);
 
-    void decreaseSoldQuantity(@Param("id") Long id, @Param("quantity") Long quantity);
+    void decreaseSoldQuantity(@Param("id") Long id, @Param("quantity") Integer quantity);
 
     void incrementViewCount(@Param("id") Long id);
